Memory Leak Fix
I tracked through tracemalloc that the link still remains, and when using opensb, each player's entry increases the memory by 58-60 megabytes. If the player exits, the memory is not cleared and increases each time. Such a small fix now really does not allow the memory to increase by 50-60 MB, but only by 200-300 KB